Revert "Testing account users"
This reverts commit 2ee7e166fb8ec91f97fec8b337c89df0522f2a0e.
