+++ src/react/utils/IAMhooks.ts
@@ -62,9 +62,11 @@ export const useAwsPaginatedEntities = <
     }
   }, [internalStatus, pageIndex]);
   useMemo(() => {
-    if (internalStatus === 'idle' || internalStatus === 'loading' || internalStatus === 'error') {
+    if (internalStatus === 'loading' || internalStatus === 'error') {
       setStatus(internalStatus);
-    } else if (internalStatus === 'success' && !hasNextPage && !isFetchingNextPage) {
+    }
+
+    if (internalStatus === 'success' && !hasNextPage && !isFetchingNextPage) {
       setStatus('success');
       setFirstPageStatus('success'); //ensure firstPageStatus is success when loading data from the cache
     } else {
